requirements.txt
flask
flask_session
flask_cors
sqlalchemy
requests
msal
identity
pillow